fix(output buffer): Fix that only the first elem was kept on device
This commit fixes the difference between host and device
implementation of myAtomicInc: The second parameter defines the
maximum after which the counter gets reset to zero. As it was set
to 1, the device implementation only kept the last element added,
resulting in silent withholding of discovered violations while
the host implementation incremented forever.