feat(grapple): Implement start over strategy
This commit introduces the new "start over" strategy: It is based
on the discovery that we need smaller hash tables to cause more
collision, resulting in a better partition of the state space
between VTs. However, only reducing the hash table's size results
in a highly lowered coverage, as each VT starts at the initial
state and can only reach as deep as the hash table can keep states.

To circumvent this problem, this commit implements "starting over":
Each VT's thread keeps a ring buffer of the last visited, unmarked
states. When the VT is done, i.e. the hash table calls every state
as already visited, it clears the hash table and each thread starts
a new state space exploration using the last visited, unmarked
states as new initial states.

This strategy results in:

1) a way deeper state space exploration, which is only limited by
   the constant defining the number of start overs
2) a way better partition among VTs, due to smaller hash hash
   tables and thus more collisions
2) less memory usage due to smaller hash tables