fix(HyperLogLog): FINALLY counting is fixed and correct!!!!!
To merge two HLLs, we just have to take the max of each register
pair. No binary OR needed.

This commit also switches to atomic operations within the device
functions of our HLL implementation, just to be save.

Last but not least, we increase the precision of the HLLs by
increasing the number of registers!