feat(grapple): Add BFS frontier inspection
This commit adds a new build parameter: -DGRAPPLE_INSPECT_HT

When it is set, only a single VT is executed. After each round
of its BFS, the number of newly visited and again visited
(= failed) states of the round are printed.

This makes it possible to observe how e.g. different models
are explored by the BFS.